WIP:Improve Performance
Limit & Stop Loss Orders are executed correctly.
Might be a delay where this execution happened, I mean if the candles that the strategy use every 1 hour candle than the order will be execute with the **right price** but at most with delay of 59 minutes.

Not supposed to effect the most interest part of th backtests, the profit of the stategy.