Bug fix - service.candle.generate_candle_from_one_minutes
This method forming from 1 minute candles a candle with bigger interval.
There are cases that might be a bug and candles will be missed in this interval (Some generated candles like test_candles_0 that there are gaps in there, or even worst the candles that have been download from an exchange has gaps.)

When this scenario happened you cant just take the next candles and form other candle than reality, instead I raise a value exception if there are gaps.