import_candles() when called via research exits early
import_candles() when called via jesse.research exits early when the start_date of the candles is before the first available date. This is due to the "return" returning nothing and calling run() on a separate line instead of returning the result of run().

I have used the existing running_via_dashboard if statement to run the current version of the code, if not running_via_dashboard then instead the recursive run() is called, passed the relevant parameters, and then returned.

This results in the expected behavior of receiving the "Successfully imported candles for..." message regardless of whether the start_date that was passed to import_candles() via jesse.research was before the first available date or not.

N.B The "No candle exists in the market for..." message is not known to the user when they call import_candles(), They only ever receive the success message. Considering the return statements in run() are only returned when not using the dashboard, i.e. via research, a named tuple could be returned to return the "No candle exists in the market for..." if there is one, and the success message. Not a deal breaker but a nice to know for the user if they are logging the candle downloading. Currently the only way to know if that occurred is to compare the start_date that was passed to import_candles() and to look at the success message and see if the dates match.