Add min_qty attribute to live trading modes in Strategy class
The 'min_qty' attribute has been added to the Strategy class as a live trading mode only attribute. It represents the minimal quantity that can be traded for a particular symbol on a given exchange. The feature has been implemented across multiple exchanges as part of the symbol dictionary returned, enhancing the precision of trade quantities.